@@ -55,4 +55,4 @@ group :production do
   gem 'rails_12factor'
 end
 
-ruby '2.0.0'
+ruby '2.1.0'
